Fix translations for tokens with multiple types
It used to be the case, when presented with a token with more than one
type, that a translation would be picked ‘arbitrarily’ from the list of
translations, so the translation may represent the incorrect type. For
example ‘sar’ could be a verb ‘use’ or a noun ‘to things’ (ay-sa-ru) and
this would be parsed correctly in both cases, but then for both the
translation ‘thing’ would be used.

Fixed now by taking the expected type into account when searching for a
translation. (Ugly but what can I do ...)